requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
greenlet==1.1.1
idna==3.2
numpy==1.21.1
pandas==1.3.1
psycopg2==2.9.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
spotipy==2.18.0
SQLAlchemy==1.4.22
urllib3==1.26.6
